requirements.txt
flask
flask-cors
pandas
joblib
requests
python-dotenv
